backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
twilio==8.10.0
supabase==2.3.0
reportlab==4.0.8
python-dotenv==1.0.0
